python/requirements.txt
absl-py==2.1.0
astunparse==1.6.3
beautifulsoup4==4.12.3
certifi==2024.12.14
charset-normalizer==3.4.1
contourpy==1.3.1
cycler==0.12.1
filelock==3.16.1
flatbuffers==24.12.23
fonttools==4.55.3
gast==0.6.0
gdown==5.2.0
google-pasta==0.2.0
grpcio==1.68.1
h5py==3.12.1
idna==3.10
imageio==2.36.1
keras==3.7.0
kiwisolver==1.4.8
lazy_loader==0.4
libclang==18.1.1
Markdown==3.7
markdown-it-py==3.0.0
MarkupSafe==3.0.2
matplotlib==3.10.0
mdurl==0.1.2
ml-dtypes==0.4.1
namex==0.0.8
networkx==3.4.2
numpy==1.26.4
opencv-python==4.10.0.84
opennsfw2==0.14.0
opt_einsum==3.4.0
optree==0.13.1
packaging==24.2
pillow==11.0.0
protobuf==4.25.5
Pygments==2.18.0
pyparsing==3.2.0
PySocks==1.7.1
python-dateutil==2.9.0.post0
requests==2.32.3
rich==13.9.4
scikit-image==0.25.0
scipy==1.14.1
six==1.17.0
soupsieve==2.6
tensorboard==2.17.1
tensorboard-data-server==0.7.2
tensorflow-cpu==2.17.0
tensorflow-io-gcs-filesystem==0.37.1
termcolor==2.5.0
tifffile==2024.12.12
tqdm==4.67.1
typing_extensions==4.12.2
urllib3==2.3.0
Werkzeug==3.1.3
wrapt==1.17.0
